+++ src/app/ControlView/Nodes/BooleanToggle.tsx
@@ -1,25 +1,31 @@
 import { Node, NodeProps } from '@xyflow/react';
 
 import NodeShell from '../NodeShell';
-import { AttributeBool } from '@/app/attribute';
 import { useAttributeBoolListener } from '../AttributeListener';
 
 import BooleanToggle from './Widgets/BooleanToggle';
 
 export type BooleanToggleNode = Node<{
-    attribute: AttributeBool;
+    attribute: {
+        name: string;
+        path: string;
+        mode: string;
+        classPath: string;
+        driver: string;
+    };
 }>;
 
 const BooleanToggleNode: React.FC<NodeProps<BooleanToggleNode>> = (props) => {
-    const { value, publish, connected, isFreshValue } = useAttributeBoolListener({
-        attribute: props.data.attribute,
+    const { publish, connected, isFreshValue, value } = useAttributeBoolListener({
+        path: props.data?.attribute.path,
+        mode: props.data?.attribute.mode,
     });
 
     return (
         <NodeShell
             topLeft={props.data?.attribute.name}
             topRight={props.data?.attribute.classPath}
-            bottomRight={props.data?.attribute.parentDriver}
+            bottomRight={props.data?.attribute.driver}
             selected={props.selected || false}
             disabled={!connected}
             animateBorder={isFreshValue}