+++ src/app/ControlView/Nodes/SiDisplay.tsx
@@ -1,32 +1,38 @@
 import { Node, NodeProps } from '@xyflow/react';
 
 import NodeShell from '../NodeShell';
-import { AttributeSi } from '@/app/attribute';
 
 import { useAttributeSiListener } from '../AttributeListener';
 import NumberDisplay from './Widgets/NumberDisplay';
 
 export type SiDisplayNode = Node<{
-    attribute: AttributeSi;
+    attribute: {
+        name: string;
+        path: string;
+        mode: string;
+        classPath: string;
+        driver: string;
+    };
 }>;
 
 const SiDisplayNode: React.FC<NodeProps<SiDisplayNode>> = (props) => {
     const { value, settings, isFreshValue, connected } = useAttributeSiListener({
-        attribute: props.data.attribute,
+        path: props.data?.attribute.path,
+        mode: props.data?.attribute.mode,
     });
 
     return (
         <NodeShell
             topLeft={props.data?.attribute.name}
             topRight={props.data?.attribute.classPath}
-            bottomRight={props.data?.attribute.parentDriver}
+            bottomRight={props.data?.attribute.driver}
             selected={props.selected || false}
             disabled={!connected}
             animateBorder={isFreshValue}
         >
             <NumberDisplay
                 value={value}
-                unit={settings.unit || ''}
+                unit={settings?.unit}
             />
         </NodeShell>
     );