+++ src/app/ControlView/Nodes/SiSpinner.tsx
@@ -1,7 +1,5 @@
 import { Node, NodeProps } from '@xyflow/react';
 
-import { AttributeSi } from '@/app/attribute';
-
 import NumberSpinner from './Widgets/NumberSpinner';
 
 import NodeShell from '../NodeShell';
@@ -9,19 +7,26 @@ import NodeShell from '../NodeShell';
 import { useAttributeSiListener } from '../AttributeListener';
 
 export type SiSpinnerNode = Node<{
-    attribute: AttributeSi;
+    attribute: {
+        name: string;
+        path: string;
+        mode: string;
+        classPath: string;
+        driver: string;
+    };
 }>;
 
 const SiSpinnerNode: React.FC<NodeProps<SiSpinnerNode>> = (props) => {
     const { value, settings, isFreshValue, connected, publish } = useAttributeSiListener({
-        attribute: props.data.attribute,
+        path: props.data?.attribute.path,
+        mode: props.data?.attribute.mode,
     });
 
     return (
         <NodeShell
             topLeft={props.data?.attribute.name}
             topRight={props.data?.attribute.classPath}
-            bottomRight={props.data?.attribute.parentDriver}
+            bottomRight={props.data?.attribute.driver}
             selected={props.selected || false}
             disabled={!connected}
             animateBorder={isFreshValue}
@@ -29,9 +34,9 @@ const SiSpinnerNode: React.FC<NodeProps<SiSpinnerNode>> = (props) => {
             <NumberSpinner
                 value={value}
                 onNewValue={publish}
-                min={settings.min}
-                max={settings.max}
-                unit={settings.unit}
+                min={settings?.min}
+                max={settings?.max}
+                unit={settings?.unit}
             />
         </NodeShell>
     );